requirements.txt
langchain-google-genai
python-dotenv
langchain-redis
redis
langchain-core
langchain-community
